+++ app/page.tsx
@@ -1,7 +1,13 @@
+"use client";
+
+import { Button } from "@/stories/Atoms/Button";
+
 export default function Home() {
   return (
     <main>
       Hello, world!
+      <Button kind="primary" label="test"></Button>
+      {/* <Header /> */}
     </main>
-  )
+  );
 }