+++ stories/Example/Button.tsx
@@ -46,11 +46,11 @@ export const Button = ({
       {...props}
     >
       {label}
-      <style jsx>{`
+      {/* <style jsx>{`
         button {
           background-color: ${backgroundColor};
         }
-      `}</style>
+      `}</style> */}
     </button>
   );
 };