+++ stories/index.ts
@@ -1,1 +1,5 @@
-export * from "./Atoms";
+export { Button } from "./Atoms/Button";
+export { Button2 } from "./Atoms/Button2";
+// export { Button } from "./Example/Button";
+export { Header } from "./Example/Header";
+export { Page } from "./Example/Page";